Split up markdown-js into multiple files, uses Grunt to produce browser distribution.
