+++ .gitignore
@@ -2,3 +2,5 @@
 .DS_Store
 /node_modules/
 /npm-debug.log
+dist
+lib/markdown.js
