+++ inc/footer-node.js
@@ -1,0 +1,8 @@
+
+	expose.Markdown = Markdown;
+	expose.parse = Markdown.parse;
+	expose.toHTML = Markdown.toHTML;
+	expose.toHTMLTree = Markdown.toHTMLTree;
+	expose.renderJsonML = Markdown.renderJsonML;
+
+})(exports);
