+++ inc/footer-web.js
@@ -1,0 +1,11 @@
+
+	expose.Markdown = Markdown;
+	expose.parse = Markdown.parse;
+	expose.toHTML = Markdown.toHTML;
+	expose.toHTMLTree = Markdown.toHTMLTree;
+	expose.renderJsonML = Markdown.renderJsonML;
+
+})(function() {
+	window.markdown = {};
+	return window.markdown;
+}());
