+++ src/markdown.js
@@ -1,0 +1,7 @@
+if (typeof define !== 'function') { var define = require('amdefine')(module) }
+
+// Include all our depndencies and return the resulting library.
+
+define(['./parser', './markdown_helpers', './render_tree', './dialects/gruber', './dialects/maruku'], function(Markdown) {
+  return Markdown;
+});
