+++ src/markdown_helpers.js
@@ -1,0 +1,87 @@
+if (typeof define !== 'function') { var define = require('amdefine')(module) }
+
+define([], function (w) {
+
+  var MarkdownHelpers = {};
+
+  // For Spidermonkey based engines
+  function mk_block_toSource() {
+    return "Markdown.mk_block( " +
+            uneval(this.toString()) +
+            ", " +
+            uneval(this.trailing) +
+            ", " +
+            uneval(this.lineNumber) +
+            " )";
+  }
+
+  // node
+  function mk_block_inspect() {
+    var util = require("util");
+    return "Markdown.mk_block( " +
+            util.inspect(this.toString()) +
+            ", " +
+            util.inspect(this.trailing) +
+            ", " +
+            util.inspect(this.lineNumber) +
+            " )";
+
+  }
+
+  MarkdownHelpers.mk_block = function(block, trail, line) {
+    // Be helpful for default case in tests.
+    if ( arguments.length === 1 )
+      trail = "\n\n";
+
+    // We actually need a String object, not a string primitive
+    /* jshint -W053 */
+    var s = new String(block);
+    s.trailing = trail;
+    // To make it clear its not just a string
+    s.inspect = mk_block_inspect;
+    s.toSource = mk_block_toSource;
+
+    if ( line !== undefined )
+      s.lineNumber = line;
+
+    return s;
+  };
+
+
+  var isArray = MarkdownHelpers.isArray = Array.isArray || function(obj) {
+    return Object.prototype.toString.call(obj) === "[object Array]";
+  };
+
+  // Don't mess with Array.prototype. Its not friendly
+  if ( Array.prototype.forEach ) {
+    MarkdownHelpers.forEach = function forEach( arr, cb, thisp ) {
+      return arr.forEach( cb, thisp );
+    };
+  }
+  else {
+    MarkdownHelpers.forEach = function forEach(arr, cb, thisp) {
+      for (var i = 0; i < arr.length; i++)
+        cb.call(thisp || arr, arr[i], i, arr);
+    };
+  }
+
+  MarkdownHelpers.isEmpty = function isEmpty( obj ) {
+    for ( var key in obj ) {
+      if ( hasOwnProperty.call( obj, key ) )
+        return false;
+    }
+    return true;
+  };
+
+  MarkdownHelpers.extract_attr = function extract_attr( jsonml ) {
+    return isArray(jsonml)
+        && jsonml.length > 1
+        && typeof jsonml[ 1 ] === "object"
+        && !( isArray(jsonml[ 1 ]) )
+        ? jsonml[ 1 ]
+        : undefined;
+  };
+
+  return MarkdownHelpers;
+
+});
