+++ src/parser.js
@@ -1,0 +1,200 @@
+if (typeof define !== 'function') { var define = require('amdefine')(module) }
+
+define(['./markdown_helpers', './core'], function(MarkdownHelpers, Markdown) {
+
+  // Imported functions
+  var mk_block = Markdown.mk_block = MarkdownHelpers.mk_block,
+      isArray = MarkdownHelpers.isArray;
+
+  /**
+   *  parse( markdown, [dialect] ) -> JsonML
+   *  - markdown (String): markdown string to parse
+   *  - dialect (String | Dialect): the dialect to use, defaults to gruber
+   *
+   *  Parse `markdown` and return a markdown document as a Markdown.JsonML tree.
+   **/
+  Markdown.parse = function( source, dialect ) {
+    // dialect will default if undefined
+    var md = new Markdown( dialect );
+    return md.toTree( source );
+  };
+
+  function count_lines( str ) {
+    var n = 0,
+        i = -1;
+    while ( ( i = str.indexOf("\n", i + 1) ) !== -1 )
+      n++;
+    return n;
+  }
+
+  // Internal - split source into rough blocks
+  Markdown.prototype.split_blocks = function splitBlocks( input ) {
+    input = input.replace(/(\r\n|\n|\r)/g, "\n");
+    // [\s\S] matches _anything_ (newline or space)
+    // [^] is equivalent but doesn't work in IEs.
+    var re = /([\s\S]+?)($|\n#|\n(?:\s*\n|$)+)/g,
+        blocks = [],
+        m;
+
+    var line_no = 1;
+
+    if ( ( m = /^(\s*\n)/.exec(input) ) !== null ) {
+      // skip (but count) leading blank lines
+      line_no += count_lines( m[0] );
+      re.lastIndex = m[0].length;
+    }
+
+    while ( ( m = re.exec(input) ) !== null ) {
+      if (m[2] === "\n#") {
+        m[2] = "\n";
+        re.lastIndex--;
+      }
+      blocks.push( mk_block( m[1], m[2], line_no ) );
+      line_no += count_lines( m[0] );
+    }
+
+    return blocks;
+  };
+
+  /**
+   *  Markdown#processBlock( block, next ) -> undefined | [ JsonML, ... ]
+   *  - block (String): the block to process
+   *  - next (Array): the following blocks
+   *
+   * Process `block` and return an array of JsonML nodes representing `block`.
+   *
+   * It does this by asking each block level function in the dialect to process
+   * the block until one can. Succesful handling is indicated by returning an
+   * array (with zero or more JsonML nodes), failure by a false value.
+   *
+   * Blocks handlers are responsible for calling [[Markdown#processInline]]
+   * themselves as appropriate.
+   *
+   * If the blocks were split incorrectly or adjacent blocks need collapsing you
+   * can adjust `next` in place using shift/splice etc.
+   *
+   * If any of this default behaviour is not right for the dialect, you can
+   * define a `__call__` method on the dialect that will get invoked to handle
+   * the block processing.
+   */
+  Markdown.prototype.processBlock = function processBlock( block, next ) {
+    var cbs = this.dialect.block,
+        ord = cbs.__order__;
+
+    if ( "__call__" in cbs )
+      return cbs.__call__.call(this, block, next);
+
+    for ( var i = 0; i < ord.length; i++ ) {
+      //D:this.debug( "Testing", ord[i] );
+      var res = cbs[ ord[i] ].call( this, block, next );
+      if ( res ) {
+        //D:this.debug("  matched");
+        if ( !isArray(res) || ( res.length > 0 && !( isArray(res[0]) ) ) )
+          this.debug(ord[i], "didn't return a proper array");
+        //D:this.debug( "" );
+        return res;
+      }
+    }
+
+    // Uhoh! no match! Should we throw an error?
+    return [];
+  };
+
+  Markdown.prototype.processInline = function processInline( block ) {
+    return this.dialect.inline.__call__.call( this, String( block ) );
+  };
+
+  /**
+   *  Markdown#toTree( source ) -> JsonML
+   *  - source (String): markdown source to parse
+   *
+   *  Parse `source` into a JsonML tree representing the markdown document.
+   **/
+  // custom_tree means set this.tree to `custom_tree` and restore old value on return
+  Markdown.prototype.toTree = function toTree( source, custom_root ) {
+    var blocks = source instanceof Array ? source : this.split_blocks( source );
+
+    // Make tree a member variable so its easier to mess with in extensions
+    var old_tree = this.tree;
+    try {
+      this.tree = custom_root || this.tree || [ "markdown" ];
+
+      blocks_loop:
+      while ( blocks.length ) {
+        var b = this.processBlock( blocks.shift(), blocks );
+
+        // Reference blocks and the like won't return any content
+        if ( !b.length )
+          continue blocks_loop;
+
+        this.tree.push.apply( this.tree, b );
+      }
+      return this.tree;
+    }
+    finally {
+      if ( custom_root )
+        this.tree = old_tree;
+    }
+  };
+
+  // Noop by default
+  Markdown.prototype.debug = function () {
+    var args = Array.prototype.slice.call( arguments);
+    args.unshift(this.debug_indent);
+    if ( typeof print !== "undefined" )
+        print.apply( print, args );
+    if ( typeof console !== "undefined" && typeof console.log !== "undefined" )
+        console.log.apply( null, args );
+  };
+
+  Markdown.prototype.loop_re_over_block = function( re, block, cb ) {
+    // Dont use /g regexps with this
+    var m,
+        b = block.valueOf();
+
+    while ( b.length && (m = re.exec(b) ) !== null ) {
+      b = b.substr( m[0].length );
+      cb.call(this, m);
+    }
+    return b;
+  };
+
+  // Build default order from insertion order.
+  Markdown.buildBlockOrder = function(d) {
+    var ord = [];
+    for ( var i in d ) {
+      if ( i === "__order__" || i === "__call__" )
+        continue;
+      ord.push( i );
+    }
+    d.__order__ = ord;
+  };
+
+  // Build patterns for inline matcher
+  Markdown.buildInlinePatterns = function(d) {
+    var patterns = [];
+
+    for ( var i in d ) {
+      // __foo__ is reserved and not a pattern
+      if ( i.match( /^__.*__$/) )
+        continue;
+      var l = i.replace( /([\\.*+?|()\[\]{}])/g, "\\$1" )
+               .replace( /\n/, "\\n" );
+      patterns.push( i.length === 1 ? l : "(?:" + l + ")" );
+    }
+
+    patterns = patterns.join("|");
+    d.__patterns__ = patterns;
+    //print("patterns:", uneval( patterns ) );
+
+    var fn = d.__call__;
+    d.__call__ = function(text, pattern) {
+      if ( pattern !== undefined )
+        return fn.call(this, text, pattern);
+      else
+        return fn.call(this, text, patterns);
+    };
+  };
+
+  return Markdown;
+});
