+++ src/render_tree.js
@@ -1,0 +1,287 @@
+if (typeof define !== 'function') { var define = require('amdefine')(module) }
+
+define(['./core', './markdown_helpers'], function(Markdown, MarkdownHelpers) {
+
+  var extract_attr = MarkdownHelpers.extract_attr;
+
+  /**
+   *  renderJsonML( jsonml[, options] ) -> String
+   *  - jsonml (Array): JsonML array to render to XML
+   *  - options (Object): options
+   *
+   *  Converts the given JsonML into well-formed XML.
+   *
+   *  The options currently understood are:
+   *
+   *  - root (Boolean): wether or not the root node should be included in the
+   *    output, or just its children. The default `false` is to not include the
+   *    root itself.
+   */
+  Markdown.renderJsonML = function( jsonml, options ) {
+    options = options || {};
+    // include the root element in the rendered output?
+    options.root = options.root || false;
+
+    var content = [];
+
+    if ( options.root ) {
+      content.push( render_tree( jsonml ) );
+    }
+    else {
+      jsonml.shift(); // get rid of the tag
+      if ( jsonml.length && typeof jsonml[ 0 ] === "object" && !( jsonml[ 0 ] instanceof Array ) )
+        jsonml.shift(); // get rid of the attributes
+
+      while ( jsonml.length )
+        content.push( render_tree( jsonml.shift() ) );
+    }
+
+    return content.join( "\n\n" );
+  };
+
+
+  /**
+   *  toHTMLTree( markdown, [dialect] ) -> JsonML
+   *  toHTMLTree( md_tree ) -> JsonML
+   *  - markdown (String): markdown string to parse
+   *  - dialect (String | Dialect): the dialect to use, defaults to gruber
+   *  - md_tree (Markdown.JsonML): parsed markdown tree
+   *
+   *  Turn markdown into HTML, represented as a JsonML tree. If a string is given
+   *  to this function, it is first parsed into a markdown tree by calling
+   *  [[parse]].
+   **/
+  Markdown.toHTMLTree = function toHTMLTree( input, dialect , options ) {
+
+    // convert string input to an MD tree
+    if ( typeof input === "string" )
+      input = this.parse( input, dialect );
+
+    // Now convert the MD tree to an HTML tree
+
+    // remove references from the tree
+    var attrs = extract_attr( input ),
+        refs = {};
+
+    if ( attrs && attrs.references )
+      refs = attrs.references;
+
+    var html = convert_tree_to_html( input, refs , options );
+    merge_text_nodes( html );
+    return html;
+  };
+
+  /**
+   *  toHTML( markdown, [dialect]  ) -> String
+   *  toHTML( md_tree ) -> String
+   *  - markdown (String): markdown string to parse
+   *  - md_tree (Markdown.JsonML): parsed markdown tree
+   *
+   *  Take markdown (either as a string or as a JsonML tree) and run it through
+   *  [[toHTMLTree]] then turn it into a well-formated HTML fragment.
+   **/
+  Markdown.toHTML = function toHTML( source , dialect , options ) {
+    var input = this.toHTMLTree( source , dialect , options );
+
+    return this.renderJsonML( input );
+  };
+
+
+  function escapeHTML( text ) {
+    return text.replace( /&/g, "&amp;" )
+               .replace( /</g, "&lt;" )
+               .replace( />/g, "&gt;" )
+               .replace( /"/g, "&quot;" )
+               .replace( /'/g, "&#39;" );
+  }
+
+  function render_tree( jsonml ) {
+    // basic case
+    if ( typeof jsonml === "string" )
+      return escapeHTML( jsonml );
+
+    var tag = jsonml.shift(),
+        attributes = {},
+        content = [];
+
+    if ( jsonml.length && typeof jsonml[ 0 ] === "object" && !( jsonml[ 0 ] instanceof Array ) )
+      attributes = jsonml.shift();
+
+    while ( jsonml.length )
+      content.push( render_tree( jsonml.shift() ) );
+
+    var tag_attrs = "";
+    for ( var a in attributes )
+      tag_attrs += " " + a + '="' + escapeHTML( attributes[ a ] ) + '"';
+
+    // be careful about adding whitespace here for inline elements
+    if ( tag === "img" || tag === "br" || tag === "hr" )
+      return "<"+ tag + tag_attrs + "/>";
+    else
+      return "<"+ tag + tag_attrs + ">" + content.join( "" ) + "</" + tag + ">";
+  }
+
+  function convert_tree_to_html( tree, references, options ) {
+    var i;
+    options = options || {};
+
+    // shallow clone
+    var jsonml = tree.slice( 0 );
+
+    if ( typeof options.preprocessTreeNode === "function" )
+      jsonml = options.preprocessTreeNode(jsonml, references);
+
+    // Clone attributes if they exist
+    var attrs = extract_attr( jsonml );
+    if ( attrs ) {
+      jsonml[ 1 ] = {};
+      for ( i in attrs ) {
+        jsonml[ 1 ][ i ] = attrs[ i ];
+      }
+      attrs = jsonml[ 1 ];
+    }
+
+    // basic case
+    if ( typeof jsonml === "string" )
+      return jsonml;
+
+    // convert this node
+    switch ( jsonml[ 0 ] ) {
+      case "header":
+        jsonml[ 0 ] = "h" + jsonml[ 1 ].level;
+        delete jsonml[ 1 ].level;
+        break;
+      case "bulletlist":
+        jsonml[ 0 ] = "ul";
+        break;
+      case "numberlist":
+        jsonml[ 0 ] = "ol";
+        break;
+      case "listitem":
+        jsonml[ 0 ] = "li";
+        break;
+      case "para":
+        jsonml[ 0 ] = "p";
+        break;
+      case "markdown":
+        jsonml[ 0 ] = "html";
+        if ( attrs )
+          delete attrs.references;
+        break;
+      case "code_block":
+        jsonml[ 0 ] = "pre";
+        i = attrs ? 2 : 1;
+        var code = [ "code" ];
+        code.push.apply( code, jsonml.splice( i, jsonml.length - i ) );
+        jsonml[ i ] = code;
+        break;
+      case "inlinecode":
+        jsonml[ 0 ] = "code";
+        break;
+      case "img":
+        jsonml[ 1 ].src = jsonml[ 1 ].href;
+        delete jsonml[ 1 ].href;
+        break;
+      case "linebreak":
+        jsonml[ 0 ] = "br";
+      break;
+      case "link":
+        jsonml[ 0 ] = "a";
+        break;
+      case "link_ref":
+        jsonml[ 0 ] = "a";
+
+        // grab this ref and clean up the attribute node
+        var ref = references[ attrs.ref ];
+
+        // if the reference exists, make the link
+        if ( ref ) {
+          delete attrs.ref;
+
+          // add in the href and title, if present
+          attrs.href = ref.href;
+          if ( ref.title )
+            attrs.title = ref.title;
+
+          // get rid of the unneeded original text
+          delete attrs.original;
+        }
+        // the reference doesn't exist, so revert to plain text
+        else {
+          return attrs.original;
+        }
+        break;
+      case "img_ref":
+        jsonml[ 0 ] = "img";
+
+        // grab this ref and clean up the attribute node
+        var ref = references[ attrs.ref ];
+
+        // if the reference exists, make the link
+        if ( ref ) {
+          delete attrs.ref;
+
+          // add in the href and title, if present
+          attrs.src = ref.href;
+          if ( ref.title )
+            attrs.title = ref.title;
+
+          // get rid of the unneeded original text
+          delete attrs.original;
+        }
+        // the reference doesn't exist, so revert to plain text
+        else {
+          return attrs.original;
+        }
+        break;
+    }
+
+    // convert all the children
+    i = 1;
+
+    // deal with the attribute node, if it exists
+    if ( attrs ) {
+      // if there are keys, skip over it
+      for ( var key in jsonml[ 1 ] ) {
+        i = 2;
+        break;
+      }
+      // if there aren't, remove it
+      if ( i === 1 )
+        jsonml.splice( i, 1 );
+    }
+
+    for ( ; i < jsonml.length; ++i ) {
+      jsonml[ i ] = convert_tree_to_html( jsonml[ i ], references, options );
+    }
+
+    return jsonml;
+  }
+
+
+  // merges adjacent text nodes into a single node
+  function merge_text_nodes( jsonml ) {
+    // skip the tag name and attribute hash
+    var i = extract_attr( jsonml ) ? 2 : 1;
+
+    while ( i < jsonml.length ) {
+      // if it's a string check the next item too
+      if ( typeof jsonml[ i ] === "string" ) {
+        if ( i + 1 < jsonml.length && typeof jsonml[ i + 1 ] === "string" ) {
+          // merge the second string into the first and remove it
+          jsonml[ i ] += jsonml.splice( i + 1, 1 )[ 0 ];
+        }
+        else {
+          ++i;
+        }
+      }
+      // if it's not a string recurse
+      else {
+        merge_text_nodes( jsonml[ i ] );
+        ++i;
+      }
+    }
+  }
+
+  return Markdown;
+});
